Update for maintainability
This should not result in *any* difference in performance or behavior
with `msgpack` handling.

This is based on the v1.1.5 branch, which Vault has permanently
hard-coded. (Which is inconsistent with the `hashicorp/raft` library
and Consul, which seems to have hard-coded v0.5.5, which appears to be a
slimmed down copy of an even older version.)

Confusingly, the `master` branch of this repo is out of sync with
`v1.1.5`.

I also added a large number of binary-compatibility tests in
`hashicorp/raft` to ensure that this library can still decode old
messages.

A few more details:

* Fix benchmarks
* Update README
* Remove all codecs other than `msgpack`
* Remove support for all versions of Go before 1.11

If we want to go forward with this, I'd recommend:

* [ ] Testing this for compatibility API and binary compatibility with
  with the v0.5.5 tag, or Consul, or both.
* [ ] Remove `fastpath` and `unsafe` versions. These are currently in
  use by Vault but not Consul, but I believe they are too dangerous to
  keep. These were removed in the v0.5.5 tag.
* [ ] Moving the current `master` branch to somewhere, perhaps
  `main-archived`.
* [ ] Moving this branch to `main`
* [ ] Cutting a `v2.0.0` release
* [ ] Update `raft` to point to v2.0.0
* [ ] Update Consul to point to v2.0.0
* [ ] Update Vault to point to v2.0.0