Add raft tests; backwards compatible time.Time
We are concerned with binary backwards compatibility for this
project, especially as the upstream has a history of making breaking
changes.

With that in mind, I collected a bunch of data from
https://github.com/hashicorp/raft for v0.5.5, v1.1.5, and v1.1.6
of `go-msgpack` to ensure that, moving forward, we can always decode
old raft logs.

For the tests to pass, we have to support decoding `time.Time`
using the `encoding/binary` marshalling format, which was used
in certain v1.1.6.

Also added in the `go.sum` file, which is needed to build with
newer versions of Go.